requirements.txt
aiofiles
aiohttp
gigachat
Markdown
openai
pyTelegramBotAPI
langchain-gigachat
python-dotenv
aioredis